cleanup

+++ src/main/groovy/com/stewsters/dusk/core/component/ai/AdvancedStats.groovy
@@ -21,9 +21,9 @@ class AdvancedStats extends BaseAi implements Ai {
     }
 
     AdvancedStats(Map params) {
-        morale = params?.morale ?: 0.5f
-        chargeProbability = params?.chargeProbability ?: 0.5f
-        retreatProbability = params?.retreatProbability ?: 0.5f
+        morale = params?.morale ?: 0.5f as float
+        chargeProbability = params?.chargeProbability ?: 0.5f as float
+        retreatProbability = params?.retreatProbability ?: 0.5f as float
 
     }
 
@@ -128,7 +128,7 @@ class AdvancedStats extends BaseAi implements Ai {
     boolean canMoveToward(Entity target) {
         int dx = target.x - entity.x
         int dy = target.y - entity.y
-        float distance = Math.sqrt(dx**2 + dy**2)
+        float distance = (float) Math.sqrt(dx**2 + dy**2)
         dx = (int) Math.round(dx / distance)
         dy = (int) Math.round(dy / distance)
